Updated memcached installs (select daemon mode). Install memcached when it's not installed and we're compiling libmemcached. Bumped version of pylibmc to be compatible with new libmemcached.

+++ gab/install.py
@@ -244,28 +244,34 @@ def install_systools():
     install('htop', 'iotop', 'sysstat', 'nethogs')
 
 
-def install_memcached():
+def install_memcached(daemon=False):
     'Install memcached server'
-    install('libevent-dev')
-    run('mkdir -p src')
-    with cd('src'):
-        run('wget http://memcached.googlecode.com/files/memcached-1.4.5.tar.gz')
-        run('tar xf memcached-1.4.5.tar.gz')
-        with cd('memcached-1.4.5'):
-            args = ['--prefix=', '--exec-prefix=/usr', '--datarootdir=/usr']
-            if getattr(env, 'is_64bit', False):
-                args.append('--enable-64bit')
-            run('./configure %s' % ' '.join(args))
-            run('make')
-            sudo('make install')
-            sudo('cp scripts/memcached-init /etc/init.d/memcached')
-            sudo('mkdir -p /usr/share/memcached')
-            sudo('cp -R scripts /usr/share/memcached')
-    install_memcached_client()
+    if not exists('/usr/bin/memcached'):
+        install('libevent-dev', 'build-essential')
+        run('mkdir -p src')
+        with cd('src'):
+            run('wget http://memcached.googlecode.com/files/memcached-1.4.5.tar.gz')
+            run('tar xf memcached-1.4.5.tar.gz')
+            with cd('memcached-1.4.5'):
+                args = ['--prefix=', '--exec-prefix=/usr', '--datarootdir=/usr']
+                if getattr(env, 'is_64bit', False):
+                    args.append('--enable-64bit')
+                run('./configure %s' % ' '.join(args))
+                run('make')
+                sudo('make install')
+                sudo('mkdir -p /usr/share/memcached')
+                sudo('cp -R scripts /usr/share/memcached')
+
+    if daemon:
+        sudo('cp /usr/share/memcached/scripts/memcached-init /etc/init.d/memcached')
+        sudo('update-rc.d memcached defaults')
+        start('memcached')
 
 
 def install_memcached_client():
     'Install libmemcached as client library for memcached'
+    if not exists('/usr/bin/memcached'):
+        install_memcached()
     install('libevent-dev', 'build-essential')
     run('mkdir -p src')
     with cd('src'):
@@ -287,10 +293,10 @@ def install_memcached_client_python():
     install('python', 'python-setuptools', 'python-dev', 'build-essential')
     run('mkdir -p src')
     with cd('src'):
-        run('wget http://pypi.python.org/packages/source/p/pylibmc/pylibmc-1.0.tar.gz#md5=182d95a6d493f3cbb7bc83ae7e275816')
-        run('tar xf pylibmc-1.0.tar.gz')
-        with cd('pylibmc-1.0'):
-            if hasattr(env, 'virtual_env'):
+        run('wget http://pypi.python.org/packages/source/p/pylibmc/pylibmc-1.1.1.tar.gz#md5=e43c54e285f8d937a3f1a916256ecc85')
+        run('tar xf pylibmc-1.1.1.tar.gz')
+        with cd('pylibmc-1.1.1'):
+            if hasattr(env, 'virtual_env') and exists(env.virtual_env):
                 run('. %(virtual_env)s/bin/activate; python setup.py install --with-libmemcached=/usr/local/lib' % env)
             else:
                 sudo('python setup.py install --with-libmemcached=/usr/local/lib' % env)
